ngrammatic

Struct CorpusBuilder

pub struct CorpusBuilder { /* private fields */ }

Build an Ngram Corpus, one setting at a time.

Implementations

impl CorpusBuilder

pub fn new() -> Self

Initialize a new instance of an CorpusBuilder, with a default arity of 2, padding set to Auto, for the given texts. The default key_trans function is a pass-through, leaving the keys unmodified.

let mut corpus = CorpusBuilder::new().finish();
corpus.add_text("tomato");
let results = corpus.search("tomacco", 0.40);
if let Some(result) = results.first() {
    println!("Closest match to 'tomacco' in the corpus was {}", result.text);
} else {
    println!("The corpus contained no words similar to 'tomacco'.");
}

pub fn pad_left(self, pad_left: Pad) -> Self

Set the left padding to build into the Corpus.

pub fn pad_right(self, pad_right: Pad) -> Self

Set the right padding to build into the Corpus.

pub fn pad_full(self, pad: Pad) -> Self

Set both the left and right padding to build into the Corpus.

pub fn arity(self, arity: usize) -> Self

Set arity (the n in ngram) to use for the resulting Corpus.

pub fn fill<It>(self, iterable: It) -> Self
where It: IntoIterator, It::Item: Into<String>,

Provide an iterator that will yield strings to be added to the Corpus.

pub fn key_trans( self, key_trans: Box<dyn Fn(&str) -> String + Send + Sync>, ) -> Self

A key transformation function, supplied as a boxed Fn that takes a &str and returns a String, applied to all strings that will be added to the Corpus. Searches on the Corpus will be similarly transformed.

use ngrammatic::CorpusBuilder;
let mut corpus = CorpusBuilder::new().key_trans(Box::new(|x| x.to_lowercase())).finish();
corpus.add_text("tomato");
let results = corpus.search("ToMaTo", 0.90);
if let Some(result) = results.first() {
    println!("Closest match to 'ToMaTo' in the corpus was {}", result.text);
} else {
    println!("The corpus contained no words similar to 'ToMaTo'.");
}

pub fn case_insensitive(self) -> Self

Convenience function that calls key_trans with a closure that lowercases all keys added to the Corpus.

let mut corpus = CorpusBuilder::new().case_insensitive().finish();
corpus.add_text("tomato");
let results = corpus.search("ToMaTo", 0.90);
if let Some(result) = results.first() {
    println!("Closest match to 'ToMaTo' in the corpus was {}", result.text);
} else {
    println!("The corpus contained no words similar to 'ToMaTo'.");
}

pub fn finish(self) -> Corpus

Yield a Corpus instance with all the properties set with this builder.