ngrammatic

Struct NgramBuilder

pub struct NgramBuilder { /* private fields */ }

Build an Ngram, one setting at a time.

Implementations

impl NgramBuilder

pub fn new(text: &str) -> Self

Initialize a new instance of an NgramBuilder, with a default arity of 2, padding set to Auto, for the given text.

let a = NgramBuilder::new("tomato").arity(2).finish();
if a.contains("to") {
    println!("{} contains the bigram 'to'!", a.text);
}

pub fn pad_left(self, pad_left: Pad) -> Self

Set the left padding to build into the Ngram.

let a = NgramBuilder::new("tomato").arity(2).pad_left(Pad::Pad(" ".to_string())).finish();
if a.contains(" t") {
    println!("{}, when padded, contains the bigram ' t'!", a.text);
}

pub fn pad_right(self, pad_right: Pad) -> Self

Set the right padding to build into the Ngram.

let a = NgramBuilder::new("tomato").arity(2).pad_right(Pad::Pad(" ".to_string())).finish();
if a.contains("o ") {
    println!("{}, when padded, contains the bigram 'o '!", a.text);
}

pub fn pad_full(self, pad: Pad) -> Self

Set both the left and right padding to build into the Ngram.

let a = NgramBuilder::new("tomato").arity(2).pad_full(Pad::Pad(" ".to_string())).finish();
if a.contains(" t") {
    println!("{}, when padded, contains the bigram ' t'!", a.text);
}
if a.contains("o ") {
    println!("{}, when padded, contains the bigram 'o '!", a.text);
}

pub fn arity(self, arity: usize) -> Self

Set arity (the n in ngram) to use for the resulting Ngram.

let a = NgramBuilder::new("tomato").arity(3).finish();
if a.contains("tom") {
    println!("{} contains the trigram 'tom'!", a.text);
}

pub fn finish(self) -> Ngram

Yield an Ngram instance with all the properties set with this builder.

let a = NgramBuilder::new("tomato").arity(3).finish();
if a.contains("tom") {
    println!("{} contains the trigram 'tom'!", a.text);
}